\encoding{UTF-8}
\name{get_book_comments}
\alias{get_book_comments}
\title{获取豆瓣图书的评论文章}
\description{获取豆瓣图书的评论文章.}
\usage{get_book_comments(bookid,n=100,verbose=TRUE,...)
}
\arguments{
  \item{bookid}{ 豆瓣书籍的id号,一串数字,如20326626}
  \item{n}{获取评论的数量,默认n=100}
  \item{verbose}{若\code{verbose=TRUE则显示详细过程信息，否则不显示，默认\code{TRUE}}}
}
\value{一个列表list, 包括:
  \item{book_title}{书籍名称} 
  \item{comments_amount}{评论文章总数}
  \item{comment_info}{一个data.frame,包含9列:
    \code{title}评论文章名称,
    \code{comment}评论文章内容,
    \code{time}评论发表时间,
    \code{nickname}作者昵称, 
    \code{rating}作者评分,
    \code{useful}该评论被赞为有用的数量,
    \code{unuseful}该评论被批评为无用的数量,
    \code{author_url}作者主页, 
    \code{comment_url}该评论文章的主页}
}
\note{若实际短评数小于n，则获取的是实际数，若实际数，则获取到的数量是\code{ceiling(n/20)*20}.}
\author{qxde01<\email{qxde01@@gmail.com}>
}
\seealso{
 \code{\link{get_book_notes}}
}
\examples{
## http://book.douban.com/subject/1291204/reviews
\dontrun{get_book_comments(bookid='2120492',n=50)}
}
\keyword{comment}
\keyword{review}
\keyword{douban}
